% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binaryLogLikelihood}
\alias{binaryLogLikelihood}
\title{Calculate the log-likelihood for the MLE among all binary distributions.}
\usage{
binaryLogLikelihood(intData, d)
}
\arguments{
\item{intData}{NumericVector containing the data encoded as integers.}

\item{d}{Integer specifying the number of binary variables.}
}
\value{
\code{binaryLogLikelihood} returns a numeic value with the
log-likelihood.
}
\description{
\code{binaryLogLikelihood} calculates the log-likelihood for a data set
from d binary variables in the maximum likelihood estimator among all
binary distributions.
}
\details{
The data must be given as a vector of integers, where each entry
corresponds to an observation from the d binary variables with the value of
its binary representation equal to the value of the integer.
}
\examples{
1+1

}
\author{
Kim Daniel Jakobsen
}
