% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateCondition}
\alias{calculateCondition}
\title{Calculate the distances between M and Xi}
\usage{
calculateCondition(eHatOmitNA, m, xi)
}
\arguments{
\item{eHatOmitNA}{NumericMatrix containing in the rows the indices for the entries in m and xi.}

\item{m}{NumericMatrix containing a sample second moment.}

\item{xi}{NumericMatrix to be compared with m.}
}
\value{
\code{calculateCondition} returns a vector with the distance between entries in m and xi for those entries specified by eHatOmitNA.
}
\description{
\code{calculateCondition} calculates the absolute differences between entries in the matrices M and Xi given by eHatOmitNA. This function is intended to be used by IsingMLE and IsingMLEmtp2.
}
