% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateCondition2}
\alias{calculateCondition2}
\title{Calculate the differences between M and Xi}
\usage{
calculateCondition2(e, m, xi)
}
\arguments{
\item{e}{NumericMatrix containing in the rows the indices for the entries
in m and xi.}

\item{m}{NumericMatrix containing a sample second moment.}

\item{xi}{NumericMatrix to be compared with m.}
}
\value{
\code{calculateCondition2} returns a vector with the difference
between entries in m and xi for those entries specified by e.
}
\description{
\code{calculateCondition2} calculates the differences between entries in
the matrices M and Xi given by E. This function is intended to be used by
IsingMLE and IsingMLEmtp2.
}
