% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateEmpirical}
\alias{calculateEmpirical}
\title{Calculate empirical marginal distributions of pairs in ePlus}
\usage{
calculateEmpirical(ePlus, m, xBar)
}
\arguments{
\item{ePlus}{Matrix containing pairs of variables in the rows.}

\item{m}{Matrix containing the sample second moment.}

\item{xBar}{Numeric vector containing the sample first moment.}
}
\value{
\code{CalculateEmpirical} returns a List containing a matrix with the calculated empirical distributions and an integer which is 1 of any of the empirical distributions contain zeroes and is 0 otherwise.
}
\description{
\code{CalculateEmpirical} calculates the empirical distribution of pairs of binary variables specified by ePlus from the sample first and second moments.
}
\details{
ePlus should contain the variable pairs for which the empirical distribution is to be calculated. They should be coded as integers from 1 up to the number of binary variables.
The sample moments should be ordered according to this coding, so the sample mean of variable 1 should have index 1 in xBar (using 1-indexing).
}
