% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateH}
\alias{calculateH}
\title{Calculate the natural parameter h in the Ising model from the distribution p}
\usage{
calculateH(d, p)
}
\arguments{
\item{d}{Integer specifying the number of binary variables in the sample.}

\item{p}{NumericVector specifying the distribution of the d binary
variables.}
}
\value{
\code{calculateH} returns a NumericVector with the natural
parameter h.
}
\description{
\code{calculateH} calculates the natural parameter h in the Ising model
from the distribution p.
}
\examples{
1+1

}
\author{
Kim Daniel Jakobsen
}
