% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateM}
\alias{calculateM}
\title{Calculate the sample second moment}
\usage{
calculateM(data, length, d)
}
\arguments{
\item{data}{Matrix containing samples from d binary variables in the rows.}

\item{length}{Integer specifying the number of samples in the data set.}

\item{d}{Integer specifying the number of binary variables.}
}
\value{
\code{calculateM} returns a matrix with the sample second moment.
}
\description{
\code{calculateM} calculates the sample second moment for a data set with
samples from d binary variables.
}
