% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateNewMu}
\alias{calculateNewMu}
\title{Calculate the mean vector from the distribution}
\usage{
calculateNewMu(p, d)
}
\arguments{
\item{p}{Numeric vector specifying the distribution of d binary variables.}

\item{d}{Integer specifying the number of binary variables.}
}
\value{
\code{calculateNewMu} returns a numeric vector with the means.
}
\description{
\code{calculateNewMu} calculates the mean for each binary variable from the
distribution of d binary variables.
}
