% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateNewPAndEHat}
\alias{calculateNewPAndEHat}
\title{Updates the distribution of the Ising model with MTP2 constraints and the
fitted graph}
\usage{
calculateNewPAndEHat(ePlus, d, e, p, eHat, capitalJ, iter)
}
\arguments{
\item{ePlus}{Matrix containing the edges of G to update p over.}

\item{d}{Integer containing the number of binary variables.}

\item{e}{Matrix containing the empirical distributions for the edges in
ePlus.}

\item{p}{Numeric vector containing the distribution to be updated.}

\item{eHat}{Matrix containing the edges in the independence graph of the
distribution p.}

\item{capitalJ}{Matrix containing the canonical parameter J.}

\item{iter}{Integer specifying the iteration of the algorithm.}
}
\value{
\code{calculateNewPAndEHat} returns a list containing the updated
distribution, its independence graph, and the canonical parameter J.
}
\description{
\code{calculateNewPAndEHat} is intended to be used by \code{IsingMLEmtp2}.
It updates the distribution p for the edges in ePlus when the empirical
distributions don't contain zeroes.
}
\details{
\code{calculateNewPAndEHat} doesn't use the value of eHat. It fills the
matrix eHat with the edges of the independence graph after the update.\cr
iter is used to indicate if the provided J should be used or not. If
iter = 0 J is calculated and otherwise the provided J is used.
}
\examples{
1+1

}
\author{
Kim Daniel Jakobsen
}
