% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateNewPAndEHatBoundary}
\alias{calculateNewPAndEHatBoundary}
\title{Updates the distribution of the Ising model with MTP2 constraints on the boundary}
\usage{
calculateNewPAndEHatBoundary(ePlus, d, e, p, eHat)
}
\arguments{
\item{ePlus}{Matrix containing the edges of G to update p over.}

\item{d}{Integer containing the number of binary variables.}

\item{e}{Matrix containing the empirical distributions for the edges in ePlus.}

\item{p}{Numeric vector containing the distribution to be updated.}

\item{eHat}{Matrix containing the edges in the independence graph of the distribution p.}
}
\value{
\code{calculateNewPAndEHatBoundary} returns a list containing the updated distribution and its independence graph.
}
\description{
\code{calculateNewPAndEHatBoundary} is intended to be used by \code{IsingMLEmtp2}. It updates the distribution p for the edges in ePlus when some of the empirical distributions contain zeroes.
}
