% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateNewPAndEHatBoundaryCheckLikelihood}
\alias{calculateNewPAndEHatBoundaryCheckLikelihood}
\title{Updates the distribution of the Ising model with MTP2 constraints on the boundary}
\usage{
calculateNewPAndEHatBoundaryCheckLikelihood(
  ePlus,
  d,
  e,
  p,
  eHat,
  data,
  logLikelihood
)
}
\arguments{
\item{ePlus}{Matrix containing the edges of G to update p over.}

\item{d}{Integer containing the number of binary variables.}

\item{e}{Matrix containing the empirical distributions for the edges in ePlus.}

\item{p}{Numeric vector containing the distribution to be updated.}

\item{eHat}{Matrix containing the edges in the independence graph of the distribution p.}

\item{data}{Matrix containing dataset to calculate log-likelihood.}

\item{logLikelihood}{numeric value giving the likelihood for the current p.}
}
\value{
\code{calculateNewPAndEHatBoundaryCheckLikelihood} returns a list containing the updated distribution, its independence graph, and the log-likelihood.
}
\description{
\code{calculateNewPAndEHatBoundaryCheckLikelihood} is intended to be used by \code{IsingMLEmtp2CheckLikelihood}. It updates the distribution p for the edges in ePlus when some of the empirical distributions contain zeroes and prints the likelihood after each iteration.
}
