% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateNewPAndEHatnoMTP2}
\alias{calculateNewPAndEHatnoMTP2}
\title{Updates the distribution of the Ising model and the fitted graph}
\usage{
calculateNewPAndEHatnoMTP2(ePlus, d, e, p, eHat, capitalJ)
}
\arguments{
\item{ePlus}{NumericMatrix containing the edges of G to update p over.}

\item{d}{Integer containing the number of binary variables.}

\item{e}{NumericMatrix containing the empirical distributions for the edges
in ePlus.}

\item{p}{NumericVector containing the distribution to be updated.}

\item{eHat}{NumericMatrix containing the edges in the independence graph of
the distribution p.}

\item{capitalJ}{NumericMatrix containing the canonical parameter J.}
}
\value{
\code{calculateNewPAndEHatnoMTP2} returns a list containing the
updated distribution, its independence graph, and the canonical parameter J.
}
\description{
\code{calculateNewPAndEHatnoMTP2} is intended to be used by \code{IsingMLE}.
It updates the distribution p for the edges in ePlus when the empirical
distributions don't contain zeroes.
}
\details{
\code{calculateNewPAndEHatnoMTP2} doesn't use the values of capitalJ or
eHat. It fills the matrix capitalJ with the values of each J_ij after the
update and fills the matrix eHat with the edges of the independence graph
after the update.
}
