% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateNewPAndEHatnoMTP2Boundary}
\alias{calculateNewPAndEHatnoMTP2Boundary}
\title{Updates the distribution of the Ising model on the boundary}
\usage{
calculateNewPAndEHatnoMTP2Boundary(ePlus, d, e, p)
}
\arguments{
\item{ePlus}{Matrix containing the edges of G to update p over.}

\item{d}{Integer containing the number of binary variables.}

\item{e}{Matrix containing the empirical distributions for the edges in ePlus.}

\item{p}{Numeric vector containing the distribution to be updated.}

\item{eHat}{Matrix containing the edges in the independence graph of the distribution p.}
}
\value{
\code{calculateNewPAndEHatnoMTP2Boundary} returns a list containing the updated distribution.
}
\description{
\code{calculateNewPAndEHatnoMTP2Boundary} is intended to be used by \code{IsingMLE}. It updates the distribution p for the edges in ePlus when some of the empirical distributions contain zeroes.
}
