% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateNewXi}
\alias{calculateNewXi}
\title{Calculate the second moment from the distribution}
\usage{
calculateNewXi(p, d)
}
\arguments{
\item{p}{Numeric vector specifying the distribution of d binary variables.}

\item{d}{Integer specifying the number of binary variables.}
}
\value{
\code{calculateNewXi} returns a numeric d by d matrix with the
second moments.
}
\description{
\code{calculateNewXi} takes the distribution of d binary variables and
returns the matrix of second moments.
}
