% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculatexBar}
\alias{calculatexBar}
\title{Calculate the sample first moment}
\usage{
calculatexBar(data, length, d)
}
\arguments{
\item{data}{Matrix containing samples from d binary variables in the rows.}

\item{length}{Integer specifying the number of samples in the data set.}

\item{d}{Integer specifying the number of binary variables.}
}
\value{
\code{calculatexBar} returns a numeric vector with the sample first
moment.
}
\description{
\code{calculatexBar} calculates the sample first moment for a data set with
samples from d binary variables.
}
