% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createEPlus}
\alias{createEPlus}
\title{Creates the matrix EPlus}
\usage{
createEPlus(e, m, xBar)
}
\arguments{
\item{e}{Matrix containing the edges of the graph G.}

\item{m}{Matrix containing the sample second moment.}

\item{xBar}{Numeric vector containing the sample first moment.}
}
\value{
\code{createEPlus} returns the matrix EPlus.
}
\description{
\code{createEPlus} is intended to be used with \code{isingMLE} and
\code{isingMLEmtp2}. It creates the matrix EPlus of edges uv where
\eqn{M_uv > \bar{x}_u \bar{x}_v}.
}
\examples{
1+1

}
\author{
Kim Daniel Jakobsen
}
