% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntToSign.R
\name{IntToSign}
\alias{IntToSign}
\title{Internal function. Convert integer to binary with 0 represented with -1 and 1
represented with 1.}
\usage{
IntToSign(x, digits)
}
\arguments{
\item{x}{Integer to be converted.}

\item{digits}{Integer specifying how many digits should be returned.}
}
\value{
\code{IntToSign} returns a vector with length equal to digits,
containing the sign representation of the input integer.
}
\description{
\code{IntToSign} converts an integer to a binary number where 0 is
represented with -1 and 1 is represented with 1.
}
\examples{
1+1

}
