% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsingLikelihood.R
\name{IsingLogLikelihood}
\alias{IsingLogLikelihood}
\title{Calculate log-likelihood in Ising models}
\usage{
IsingLogLikelihood(data, h = NULL, J = NULL, p = NULL)
}
\arguments{
\item{data}{Matrix or data frame containing samples from d binary variables.}

\item{h}{Vector containing the canonical parameters h.}

\item{J}{Matrix containing the canonical parameters J.}

\item{p}{Vector containing the distribution of the Ising model.}
}
\value{
\code{IsingLogLikelihood} returns a numeric value with the log-likelihood of the data set for the specified Ising model.
}
\description{
\code{IsingLogLikelihood} calculates the log-likelihood of a data set for a specified Ising model.
}
\details{
The function must be provided with either the natural parameters h and J or the distribution p. \cr
A probability in the probability p belongs to the observation whose binary value matches the value of the index (using zero-indexing).
}
