% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isingLogLikelihoodhJ}
\alias{isingLogLikelihoodhJ}
\title{Calculate loglikelihood for data set given canonical parameters}
\usage{
isingLogLikelihoodhJ(data, h, J)
}
\arguments{
\item{data}{Matrix containing samples from d binary variables in the rows.}

\item{h}{Numeric vector specifying the canonical parameter h.}

\item{J}{Matrix specifying the canonical parameter J.}
}
\value{
\code{isingLogLikelihoodhJ} returns the numeric value of the
loglikelihood function.
}
\description{
\code{isingLogLikelihoodhJ} calculates the value of the loglikelihood
function for a given data set in the natural parameters h and J.
}
\details{
The loglikelihood is calculated using the formula
\eqn{L(h,J|data) = \sum _{x} e^(h^T * x + x^T * J * x / 2 - A(h,J))},
where A(h,J) is a normalisation constant.
}
