% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isingLogLikelihoodProbs}
\alias{isingLogLikelihoodProbs}
\title{Calculate loglikelihood for data set given distribution}
\usage{
isingLogLikelihoodProbs(data, p)
}
\arguments{
\item{data}{Matrix containing samples from d binary variables in the rows.}

\item{p}{Numeric vector specifying the distribution of d binary variables.}
}
\value{
\code{isingLogLikelihoodProbs} returns the numeric value of the
loglikelihood function.
}
\description{
\code{isingLogLikelihoodProbs} calculates the value of the loglikelihood
function for a given data set in the distribution p.
}
\examples{
1+1

}
\author{
Kim Daniel Jakobsen
}
