% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsingMLEmtp2CheckLikelihood.R
\name{IsingMLEmtp2CheckLikelihood}
\alias{IsingMLEmtp2CheckLikelihood}
\title{Print the log-likelihood in each iteration.}
\usage{
IsingMLEmtp2CheckLikelihood(
  G,
  xBar = NULL,
  M = NULL,
  data = NULL,
  epsilon = 1e-04,
  epsilon2 = 0,
  maxIter = 100L,
  zeroReplace = FALSE,
  ReplaceValue = 1e-10
)
}
\arguments{
\item{G}{A list containing a vector with vertices and a matrix containing
edges in the rows.}

\item{xBar}{An optional vector containing the sample first moment of the
data. Must be specified if data is unspecified.}

\item{M}{An optional matrix containing the sample second moment of the data.
Must be specified if data is unspecified.}

\item{data}{An optional matrix or data frame containing samples from d
binary variables with outcomes coded as -1 and 1.}

\item{epsilon}{A numeric value > 0 specifying the tolerance for when the
algorithm is considered to have converged using the KKT conditions.}

\item{epsilon2}{A numeric value >= 0 specifying the tolerance for when the
algorithm is considered to have converged using the likelihood.}

\item{maxIter}{An integer specifying the maximum number of iterations to run
the algorithm.}

\item{zeroReplace}{A boolean value indicating whether or not to replace
zeroes in the empirical distribution.}

\item{ReplaceValue}{A numeric value >0 specifying which value to replace
zeroes with.}
}
\value{
\code{IsingMLEmtp2CheckLikelihood} returns a list with the estimated
distribution, estimated graph, estimated parameters, and number of iterations
until the algorithm converged.
}
\description{
\code{IsingMLEmtp2CheckLikelihood} fits graphical Ising models under MTP_2
constraints using an IPS-type algorithm and prints the log-likelihood for
each iteration.
}
\details{
The graph G specifies which undirected graph the maximum likelihood estimator
should be Markov with respect to. The graph should be specified as a list
with the first element being a vector containing the vertices of the graph
and the second element being a matrix containing the edges in the rows. Make
sure the naming of the vertices and the edges match.\cr If the maximum
likelihood estimator is on the boundary of the parameter space only the mean
value parameters will be returned. Otherwise the canonical parameters h and
J will also be returned.\cr A probability in the probability vector returned
by the function belongs to the observation whose binary value matches the
value of the index (using zero-indexing).
}
\examples{
1+1

}
\author{
Kim Daniel Jakobsen
}
