% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsingSampler.R
\name{IsingSampler}
\alias{IsingSampler}
\title{IsingSampler <to be documented>}
\usage{
IsingSampler(
  h = NULL,
  J = NULL,
  p = NULL,
  N = 1000,
  obs = TRUE,
  int = FALSE,
  matrix = FALSE
)
}
\arguments{
\item{h}{<to be documented>}

\item{J}{<to be documented>}

\item{p}{<to be documented>}

\item{N}{<to be documented>}

\item{obs}{<to be documented>}

\item{int}{<to be documented>}

\item{matrix}{<to be documented>}
}
\value{
<to be documented>
}
\description{
IsingSampler <to be documented>
}
