% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createP}
\alias{createP}
\title{Create initial distribution for MLE algorithm}
\usage{
createP(mu, d)
}
\arguments{
\item{mu}{Numeric vector specifying the mean of the distribution.}

\item{d}{Integer specifying the number of binary variables.}
}
\value{
\code{createP} returns a numeric vector with the distribution p.
}
\description{
\code{createP} is intended to be used with \code{isingMLE} and
\code{isingMLEmtp2}. It creates the vector p with the initial distribution
that the algorithms start from.
}
\details{
The initial distribution has the form
\eqn{p(x) = 2^{-|V|}\prod_{v \in V}(1 + x_v \mu_v)}.
}
\examples{
1+1

}
\author{
Kim Daniel Jakobsen
}
