% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsingSampler.R
\name{IntToSign}
\alias{IntToSign}
\title{IntToSign <to be documented>}
\usage{
IntToSign(x, dig)
}
\arguments{
\item{x}{<to be documented>}

\item{dig}{<to be documented>}
}
\value{
<to be documented>
}
\description{
IntToSign <to be documented>
}
